Renew scripts

+++ scripts/PlayerStats.cs
@@ -5,7 +5,7 @@ public class PlayerStats : MonoBehaviour {
     //Player Status
     public static int health = 0;
 
-    public static double AdaptationRate = 0; //Only for HardCore World View
+    public static float AdaptationRate = 0; //Only for HardCore World View
     public static int Sanity = 0; //Only for HardCore World View
     public static int Stamina = 0; //Only for HardCore World View
     public static int Mighty = 0; //Only for HardCore World View
